Rename partial JSON benchmarks to streaming_json
