+++ .agent/AGENTS.md
@@ -50,16 +50,16 @@ cargo bench --bench streaming_parser -- --output-format bencher | rg '^test'
 # test streaming_parser_split/5000 ... bench:  604477 ns/iter (+/- 8785)
 
 # partial JSON benchmarks
-cargo bench --bench partial_json_big -- --output-format bencher | rg '^test'
+cargo bench --bench streaming_json_medium -- --output-format bencher | rg '^test'
 
 # include external implementations
-cargo bench --features comparison --bench partial_json_big -- --output-format bencher | rg '^test'
+cargo bench --features comparison --bench streaming_json_medium -- --output-format bencher | rg '^test'
 ```
 
 ## Flamegraphs and line-level profiling
 
 This repository ships a GitHub Action that runs
-`cargo flamegraph --bench partial_json_big -- --bench` and uploads
+`cargo flamegraph --bench streaming_json_medium -- --bench` and uploads
 `flamegraph.svg`.  The `setup.sh` script installs `perf` so the same
 command can be run locally:
 
@@ -67,7 +67,7 @@ command can be run locally:
 cargo install flamegraph --locked
 sudo apt-get install -y linux-tools-common linux-tools-generic
 sudo bash -c 'echo 0 > /proc/sys/kernel/perf_event_paranoid'
-cargo flamegraph --package jsonmodem --bench partial_json_big -- --bench
+cargo flamegraph --package jsonmodem --bench streaming_json_medium -- --bench
 
 # Finished release [optimized] target(s) in 0.23s
 # Flamegraph written to flamegraph.svg
@@ -82,8 +82,8 @@ debug = "line-tables-only"
 ```
 
 ```bash
-RUSTFLAGS="-C force-frame-pointers=yes" cargo build --release --bench partial_json_big
-sudo perf record -F 999 --call-graph dwarf ./target/release/partial_json_big
+RUSTFLAGS="-C force-frame-pointers=yes" cargo build --release --bench streaming_json_medium
+sudo perf record -F 999 --call-graph dwarf ./target/release/streaming_json_medium
 sudo perf report -g fractal -F+srcline | head
 
 # Example output
@@ -91,6 +91,6 @@ sudo perf report -g fractal -F+srcline | head
 # 25.0% crates/jsonmodem/src/lexer.rs:87
 ```
 
-For deterministic instruction counts, `cargo profiler callgrind --release --bench partial_json_big` will emit
+For deterministic instruction counts, `cargo profiler callgrind --release --bench streaming_json_medium` will emit
 `callgrind.out.*` which can be viewed with `kcachegrind` and also prints the hottest lines directly in the
 terminal.
