+++ crates/jsonmodem/benches/streaming_json_common.rs
